Double long support bioinformatics lib Correct bin sizes Seems to agree with full length
git-svn-id: svn+ssh://cyverse.warwick.ac.uk/home/svn/TRegS_software/LiBiNorm/trunk@2217 d27ce7f2-3f32-4a89-82b7-fbde2a400052

+++ LiBiNormSrc/transcriptData.h
@@ -32,7 +32,7 @@ public:
 class transcriptDataMap : public vector<transcriptData>
 {
 public:
-	vector<int> freq;
+	vector<double> freq;
 
 	void remove_invalid_values();
 	int loadData(const string filename);
